Split settings controls copy and fields
